% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{accuracy}
\alias{accuracy}
\title{Accuracy}
\usage{
accuracy(expected, actual)
}
\arguments{
\item{expected}{a vector of logical values for the expected values}

\item{actual}{a vector of logical values for the actual values}
}
\value{
a list of accuracy parameters
}
\description{
Signal detection values for modules that classify papers as having a feature or not
}
