% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prereg.R
\name{aspredicted_info}
\alias{aspredicted_info}
\title{Retrieve info from AsPredicted by URL}
\usage{
aspredicted_info(ap_url)
}
\arguments{
\item{ap_url}{an AsPredicted URL}
}
\value{
a data frame of information
}
\description{
Retrieve info from AsPredicted by URL
}
\keyword{internal}
