% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/prereg.R
\name{aspredicted_retrieve}
\alias{aspredicted_retrieve}
\title{Retrieve info from AsPredicted by URL}
\usage{
aspredicted_retrieve(ap_url, id_col = 1)
}
\arguments{
\item{ap_url}{an AsPredicted URL, or a table containing them (e.g., as created by \code{aspredicted_links()})}

\item{id_col}{the index or name of the column that contains AsPredicted URLs, if id is a table}
}
\value{
a data frame of information
}
\description{
Retrieve info from AsPredicted by URL
}
