% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{check_p_values}
\alias{check_p_values}
\title{Check p-values}
\usage{
check_p_values(text)
}
\arguments{
\item{text}{the search table (or list of scienceverse objects)}
}
\value{
a table of p-values
}
\description{
Check p-values
}
\examples{
filename <- system.file("grobid", "incest.xml", package="papercheck")
study <- read_grobid(filename)
check_p_values(study)
}
