% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{demofile}
\alias{demofile}
\title{Get demo files}
\usage{
demofile(type = c("dir", "xml", "pdf"))
}
\arguments{
\item{type}{whether to return the path to the demo directory, or demo XML or PDF files}
}
\value{
vector of paths
}
\description{
Get demo files
}
\examples{
demofile()
demofile("xml")
demofile("pdf")
}
