% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{demofiles}
\alias{demofiles}
\title{Get demo files}
\usage{
demofiles(type = c("dir", "xml", "pdf"))
}
\arguments{
\item{type}{whether to return the path to the demo directory, or demo XML or PDF files}
}
\value{
vector of paths
}
\description{
Get demo files
}
\examples{
demofiles()
demofiles("xml")
demofiles("pdf")
}
