% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/emojis.R
\docType{data}
\name{emojis}
\alias{emojis}
\title{Emojis}
\format{
An object of class \code{list} of length 26.
}
\usage{
emojis
}
\description{
Useful emojis
}
\details{
\itemize{
\item General: "check", "star", warning", "stop", "x", "no", "thumbs_up", "thumbs_down", "info", "question"
\item Traffic Lights: "tl_green", "tl_yellow", "tl_red", "tl_info", "tl_na", "tl_fail"
\item Hearts: "red", "orange", "yellow", "green", "blue", "purple", "brown", "black", "white", "pink"
}
}
\keyword{datasets}
