% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/file_types.R
\docType{data}
\name{file_types}
\alias{file_types}
\title{File Types}
\format{
A data frame with 275 rows and 2 columns:
\describe{
\item{ext}{File extension}
\item{type}{Type of file}
}
}
\source{
\url{https://raw.githubusercontent.com/dyne/file-extension-list/refs/heads/master/pub/extensions.json}
}
\usage{
file_types
}
\description{
File types from extension, c("archive", "audio", "book", "code", "config", data", "exec", "font", "image", "slide", "stats", text", "video", "web")
}
\keyword{datasets}
