% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_grobid.R
\name{full_text_sections}
\alias{full_text_sections}
\title{Add section info to full text table}
\usage{
full_text_sections(ft)
}
\arguments{
\item{ft}{full text table}
}
\value{
a data frame of the classified full text
}
\description{
Add section info to full text table
}
\keyword{internal}
