% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_grobid.R
\name{get_authors}
\alias{get_authors}
\title{Get author info from XML}
\usage{
get_authors(xml)
}
\arguments{
\item{xml}{The grobid XML}
}
\value{
an author list
}
\description{
Get author info from XML
}
\keyword{internal}
