% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_grobid.R
\name{get_refs}
\alias{get_refs}
\title{Get references from grobid XML}
\usage{
get_refs(xml)
}
\arguments{
\item{xml}{The grobid XML}
}
\value{
a list with a data frame of referencesand a data frame of citation sentences
}
\description{
Get references from grobid XML
}
\keyword{internal}
