% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{github_languages}
\alias{github_languages}
\title{Get Languages from GitHub Repo}
\usage{
github_languages(repo)
}
\arguments{
\item{repo}{The URL of the repository (in the format "username/repo" or "https://github.com/username/repo")}
}
\value{
vector of languages
}
\description{
Get Languages from GitHub Repo
}
\examples{
\donttest{
  github_languages("scienceverse/papercheck")
}
}
