% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{github_readme}
\alias{github_readme}
\title{Get README from GitHub}
\usage{
github_readme(repo)
}
\arguments{
\item{repo}{The URL of the repository (in the format "username/repo" or "https://github.com/username/repo")}
}
\value{
a character string of the README contents
}
\description{
Get README from GitHub
}
\examples{
\donttest{
  github_readme("scienceverse/papercheck")
}
}
