% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/github.R
\name{github_repo}
\alias{github_repo}
\title{Get Short GitHub Repo Name}
\usage{
github_repo(repo)
}
\arguments{
\item{repo}{The URL of the repository (in the format "username/repo" or "https://github.com/username/repo")}
}
\value{
character string of short repo name
}
\description{
Get Short GitHub Repo Name
}
\examples{
github_repo("scienceverse/papercheck")
github_repo("https://github.com/scienceverse/papercheck/")
github_repo("https://github.com/scienceverse/papercheck.git")
}
