% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpt.R
\name{gpt}
\alias{gpt}
\title{Query ChatGPT}
\usage{
gpt(
  text,
  query,
  context = "You are a scientist",
  text_col = "text",
  group_by = "id",
  CHATGPT_KEY = Sys.getenv("CHATGPT_KEY"),
  gpt_model = "gpt-3.5-turbo-1106",
  chunk_size = 500,
  chunk_overlap = 100,
  temperature = 0,
  include_query = FALSE
)
}
\arguments{
\item{text}{The text to send to chatGPT (vector of strings, or data frame with the text in a column)}

\item{query}{The query to ask of chatGPT}

\item{context}{The context to send to chatGPT}

\item{text_col}{The name of the text column if text is a data frame}

\item{group_by}{the column(s) to group by if text is a data frame}

\item{CHATGPT_KEY}{your API key for ChatGPT}

\item{gpt_model}{the model name to pass to ChatOpenAI}

\item{chunk_size}{text chunk size for embeddings}

\item{chunk_overlap}{overlap between text chunks for embeddings}

\item{temperature}{temperature value for ChatGPT (0.0 to 2.0)}

\item{include_query}{Whether to include the query and context strings in the returned table}
}
\value{
a list of results
}
\description{
Ask ChatGPT any question you want about the text from a search_text()
}
\details{
You need to use your own ChatGPT API key. To avoid having to type it out, add it to the .Renviron file in the following format (you can use \code{usethis::edit_r_environ()} to access the .Renviron file)

CHATGPT_KEY="key_value_asdf"
}
