% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/llm.R
\name{json_expand}
\alias{json_expand}
\title{Expand a JSON column}
\usage{
json_expand(table, col = "answer")
}
\arguments{
\item{table}{the table with a column to expand}

\item{col}{the name or index of the column to expand (defaults to "answer" or the first column)}
}
\value{
the table plus the expanded columns
}
\description{
It is useful to ask an LLM to return data in JSON structured format, but can be frustrating to extract the data, especially where the LLM makes syntax mistakes. This function tries to expand a column with a JSON-formatted response into columns and deals with it gracefully (sets an 'error' column to "parsing error") if there are errors. It also fixes column data types, if possible.
}
\examples{
table <- data.frame(
  id = 1:5,
  answer = c(
    '{"number": "1", "letter": "A", "bool": true}',
    '{"number": "2", "letter": "B", "bool": "FALSE"}',
    '{"number": "3", "letter": "", "bool": null}',
    'oh no, the LLM misunderstood',
    '{"number": "5", "letter": ["E", "F"], "bool": false}'
  )
)

expanded <- json_expand(table, "answer")
expanded
}
