% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ml.R
\name{ml}
\alias{ml}
\title{Machine Learning Classifier}
\usage{
ml(
  text,
  model_dir,
  class_col = "classification",
  map = NULL,
  text_col = "text",
  return_prob = FALSE
)
}
\arguments{
\item{text}{the text to classify}

\item{model_dir}{the directory of the model}

\item{class_col}{the name of the classification column}

\item{map}{a vector of keys and values (e.g., \verb{c("0" = "no", "1" = yes")})}

\item{text_col}{the name of the text column if text is a data frame}

\item{return_prob}{whether to return the probability along with the classification}
}
\value{
a data frame with the text and classifications
}
\description{
Machine Learning Classifier
}
