% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{module_help}
\alias{module_help}
\title{Get Module Help}
\usage{
module_help(module = NULL)
}
\arguments{
\item{module}{the name of a module or path to a module}
}
\value{
the help text
}
\description{
See the help files for a module by name (get a list of names from \code{module_list()})
}
\examples{
module_help("marginal")
}
