% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{module_list}
\alias{module_list}
\title{List modules}
\usage{
module_list(module_dir = system.file("modules", package = "papercheck"))
}
\arguments{
\item{module_dir}{the directory to search for modules (defaults to the built-in modules)}
}
\value{
a data frame of modules
}
\description{
List modules
}
\examples{
module_list()
}
