% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{module_report}
\alias{module_report}
\title{Report from module output}
\usage{
module_report(module_output)
}
\arguments{
\item{module_output}{the output of a \code{module_run()}}
}
\value{
text
}
\description{
Report from module output
}
\examples{
filename <- demoxml()
paper <- read_grobid(filename)
op <- module_run(paper, "imprecise-p")
module_report(op) |> cat()
}
