% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{module_run}
\alias{module_run}
\title{Run a module}
\usage{
module_run(study, module)
}
\arguments{
\item{study}{a study object or a list of study objects}

\item{module}{the name of a module or path to a module to run on this object}
}
\value{
a list of the returned table and report text
}
\description{
Run a module
}
