% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{module_run_ai}
\alias{module_run_ai}
\title{Run AI module}
\usage{
module_run_ai(paper, args)
}
\arguments{
\item{paper}{the scienceverse paper object (or list of object)}

\item{args}{a list of argumentsto \code{gpt()}}
}
\value{
data frame
}
\description{
Run AI module
}
\keyword{internal}
