% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{module_run_code}
\alias{module_run_code}
\title{Run code module}
\usage{
module_run_code(study, args)
}
\arguments{
\item{study}{the scienceverse study object (or list of object)}

\item{args}{a list of arguments}
}
\value{
data frame
}
\description{
Run code module
}
\keyword{internal}
