% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{module_run_ml}
\alias{module_run_ml}
\title{Run machine learning module}
\usage{
module_run_ml(study, args)
}
\arguments{
\item{study}{the scienceverse study object (or list of object)}

\item{args}{a list of arguments}
}
\value{
data frame
}
\description{
Run machine learning module
}
\keyword{internal}
