% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osf.R
\name{osf_api_calls}
\alias{osf_api_calls}
\title{Get/set the OSF API Call Count}
\usage{
osf_api_calls(calls = NULL)
}
\arguments{
\item{calls}{the number of OSF calls made since the last reset}
}
\description{
Get/set the OSF API Call Count
}
\examples{
osf_api_calls()
}
