% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osf_check.R
\name{osf_check}
\alias{osf_check}
\title{Check for OSF links}
\usage{
osf_check(text)
}
\arguments{
\item{text}{the text table}
}
\value{
a table of OSF links, their openness, and other table info
}
\description{
Check for OSF links
}
