% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osf.R
\name{osf_check_id}
\alias{osf_check_id}
\title{Check OSF IDs}
\usage{
osf_check_id(osf_id)
}
\arguments{
\item{osf_id}{a vector of OSF IDs or URLs}
}
\value{
a vector of valid IDs, with NA in place of invalid IDs
}
\description{
Check if strings are valid OSF IDs, URLs, or waterbutler IDs. Basically an improved wrapper for \code{osfr::as_id()} that returns NA for invalid IDs in a vector.
}
\examples{
osf_check_id("pngda")
osf_check_id("osf.io/pngda")
osf_check_id("https://osf.io/pngda")
osf_check_id("https://osf.io/pngda/View")
osf_check_id("https://osf .io/png da") # rogue whitespace
osf_check_id("pnda") # invalid
}
