% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osf.R
\name{osf_children}
\alias{osf_children}
\title{List Children of an OSF Component}
\usage{
osf_children(osf_id)
}
\arguments{
\item{osf_id}{an OSF ID}
}
\value{
a data frame with child info
}
\description{
List Children of an OSF Component
}
\keyword{internal}
