% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osf.R
\name{osf_contents}
\alias{osf_contents}
\title{Get contents of OSF node}
\usage{
osf_contents(id)
}
\arguments{
\item{id}{an OSF ID or URL}
}
\value{
a data frame of the contents and info about them
}
\description{
Retrieve the contents recursiveley from an OSF node
}
