% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osf.R
\name{osf_delay}
\alias{osf_delay}
\title{Set the OSF delay}
\usage{
osf_delay(delay = NULL)
}
\arguments{
\item{delay}{the number of seconds to wait between OSF calls}
}
\description{
Sometimes the OSF gets fussy if you make too many calls, so you can set a delay of a few seconds before each call. Use \code{osf_delay()} to get or set the OSF delay.
}
\examples{
osf_delay()
}
