% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osf.R
\name{osf_files}
\alias{osf_files}
\title{List Files in an OSF Component}
\usage{
osf_files(osf_id)
}
\arguments{
\item{osf_id}{an OSF ID}
}
\value{
a data frame with file info
}
\description{
List Files in an OSF Component
}
\keyword{internal}
