% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osf.R
\name{osf_links}
\alias{osf_links}
\title{Find OSF Links in Papers}
\usage{
osf_links(paper)
}
\arguments{
\item{paper}{a paper object or paperlist object}
}
\value{
a table with the OSF url in the first (text) column
}
\description{
OSF links can be tricky to find in PDFs, since they can insert spaces in odd places, and view-only links that contain a ? are often interpreted as being split across sentences. This function is our best attempt at catching and fixing them all.
}
\examples{
osf_links(psychsci)
}
