% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osf.R
\name{osf_parent_project}
\alias{osf_parent_project}
\title{Get OSF Parent Project}
\usage{
osf_parent_project(osf_id)
}
\arguments{
\item{id}{an OSF ID}
}
\value{
the ID of the parent project
}
\description{
Get OSF Parent Project
}
\keyword{internal}
