% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper.R
\name{paper}
\alias{paper}
\title{Create a paper object}
\usage{
paper(name = "Demo Paper", ...)
}
\arguments{
\item{name}{The name of the study or a file path to a PDF or grobid XML}

\item{...}{further arguments to add}
}
\value{
A study object with class scivrs_paper
}
\description{
Create a new paper object or load a paper from PDF or XML
}
\examples{

p <- paper("Demo Paper")
}
