% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/crossref.R
\name{paper_info}
\alias{paper_info}
\title{Update Paper Info}
\usage{
paper_info(paper)
}
\arguments{
\item{paper}{a papercheck paper object or list}
}
\value{
the updated paper object
}
\description{
Check OpenAlex and/or CrossRef for paper info and update the paper object accordingly to fix import problems.
}
\examples{
ps1 <- paper_info(psychsci[[1]])
}
\keyword{internal}
