% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_grobid.R
\name{paper_validate}
\alias{paper_validate}
\title{Validate Papers}
\usage{
paper_validate(paper)
}
\arguments{
\item{paper}{a paper object or a list of paper objects}
}
\value{
a list or data frame of checks
}
\description{
A quick function to help diagnose problems with imported papers. It checks if there is a title, doi, abstract, and references.
}
\examples{
paper_validate(psychsci[[1]])
paper_validate(psychsci)
}
