% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/papercheck_app.R
\name{papercheck_app}
\alias{papercheck_app}
\title{Launch Shiny App}
\usage{
papercheck_app(study = NULL, quiet = FALSE, ...)
}
\arguments{
\item{study}{optional study to load}

\item{quiet}{whether to show the debugging messages in the console}

\item{...}{arguments to pass to shiny::runApp}
}
\value{
A study object created or edited by the app
}
\description{
Create a meta-study file interactively in a shiny app that runs locally in RStudio or your web browser (recommended). It does not connect to the web at all, so your data are completely private.
}
\examples{
\dontrun{ s <- papercheck_app() }

}
