% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/paper.R
\name{paperlist}
\alias{paperlist}
\title{Create a paperlist object}
\usage{
paperlist(..., merge_duplicates = TRUE)
}
\arguments{
\item{...}{scivrs_paper objects or lists of paper objects}

\item{merge_duplicates}{if duplicates exist, merge them}
}
\value{
An object with class scivrs_paperlist
}
\description{
Create a new paperlist object from individual paper objects or lists of paper objects
}
\examples{

p1 <- psychsci[[1]]
p2 <- psychsci[[2]]
plist <- paperlist(p1, p2)

merged <- paperlist(psychsci[1:2], psychsci[2:3])
}
\keyword{internal}
