% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pdf2grobid.R
\name{pdf2grobid}
\alias{pdf2grobid}
\title{Convert a PDF to Grobid XML}
\usage{
pdf2grobid(filename, save_path = ".")
}
\arguments{
\item{filename}{path to the PDF}

\item{save_path}{directory or file path to save to; set to NULL to save to a temp file}
}
\value{
XML object
}
\description{
Convert a PDF to Grobid XML
}
