% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.ppchk_module_list}
\alias{print.ppchk_module_list}
\title{Print Module List Object}
\usage{
\method{print}{ppchk_module_list}(x, ...)
}
\arguments{
\item{x}{The ppchk_module_list object}

\item{...}{Additional parameters for print}
}
\description{
Print Module List Object
}
\keyword{internal}
