% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.ppchk_module_output}
\alias{print.ppchk_module_output}
\title{Print Module Output}
\usage{
\method{print}{ppchk_module_output}(x, ...)
}
\arguments{
\item{x}{The ppchk_module_output object}

\item{...}{Additional parameters for print}
}
\description{
Print Module Output
}
\keyword{internal}
