% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{print.ppchk_validate}
\alias{print.ppchk_validate}
\title{Print Validation List Object}
\usage{
\method{print}{ppchk_validate}(x, ...)
}
\arguments{
\item{x}{The ppchk_validate object}

\item{...}{Additional parameters for print}
}
\description{
Print Validation List Object
}
\keyword{internal}
