% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{print.scivrs_paper}
\alias{print.scivrs_paper}
\title{Print Paper Object}
\usage{
\method{print}{scivrs_paper}(x, ...)
}
\arguments{
\item{x}{The scivrs_paper list}

\item{...}{Additional parameters for print}
}
\description{
Print Paper Object
}
\keyword{internal}
