% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_grobid.R
\name{read_grobid_xml}
\alias{read_grobid_xml}
\title{Read in grobid XML}
\usage{
read_grobid_xml(filename)
}
\arguments{
\item{filename}{The path to the XML file to be read}
}
\value{
An XML object
}
\description{
Read in grobid XML
}
\keyword{internal}
