% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report.R
\name{report}
\alias{report}
\title{Create a report}
\usage{
report(
  paper,
  modules = c("imprecise-p", "marginal", "statcheck", "osf-check", "retractionwatch",
    "ref-consistency"),
  output_file = paste0(paper$name, "_report.", output_format),
  output_format = c("qmd", "html", "pdf")
)
}
\arguments{
\item{paper}{a paper object}

\item{modules}{a vector of modules to run (names for built-in modules or paths for custom modules)}

\item{output_file}{the name of the output file}

\item{output_format}{the format to create the report in}
}
\value{
the file path the report is saved to
}
\description{
Create a report
}
\examples{
\dontrun{
filename <- demoxml()
paper <- read_grobid(filename)
report(paper)
}
}
