% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/retractionwatch.R
\docType{data}
\name{retractionwatch}
\alias{retractionwatch}
\title{RetractionWatch data}
\format{
A data frame with 44784 rows and 2 columns:
\describe{
\item{doi}{Document ID}
\item{retractionwatch}{Nature of note(s)}
}
}
\source{
\url{https://api.labs.crossref.org/data/retractionwatch}
}
\usage{
retractionwatch
}
\description{
DOIs and nature of statements from the RetractionWatch database (downloaded 2024-05-23)
}
\keyword{datasets}
