% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_text.R
\name{search_text}
\alias{search_text}
\title{Search the full text}
\usage{
search_text(
  study,
  pattern = ".*",
  section = NULL,
  return = c("sentence", "paragraph", "section", "match"),
  ignore.case = TRUE,
  ...
)
}
\arguments{
\item{study}{a study object created by \code{read_grobid} or a list of study objects}

\item{pattern}{the regex pattern to search for}

\item{section}{the section(s) to search in}

\item{return}{the kind of text to return, the full sentence, paragraph, or section that the text is in, or just the (regex) match}

\item{ignore.case}{whether to ignore case when text searching}

\item{...}{additional arguments to pass to \code{grep()} and \code{regexpr()}, such as \code{fixed = TRUE}}
}
\value{
a data frame of matches
}
\description{
Search the full text of a study. Currently only works with study objects that have full text imported from grobid (e.g., using \code{read_grobid()}).
}
\examples{
filename <- system.file("grobid", "eyecolor.xml", package="papercheck")
study <- read_grobid(filename)
search_text(study, "p\\\\s*(=|<)\\\\s*[0-9\\\\.]+", return = "match")
}
