% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gpt.R
\name{set_gpt_max_calls}
\alias{set_gpt_max_calls}
\title{Set the maximum number of calls to ChatGPT}
\usage{
set_gpt_max_calls(n = 10)
}
\arguments{
\item{n}{The maximum number of calls to ChatGPT that the gpt() function can make}
}
\description{
Set the maximum number of calls to ChatGPT
}
