% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{site_down}
\alias{site_down}
\title{Check if site is available}
\usage{
site_down(url, msg = "The website \%s is not available", error = TRUE)
}
\arguments{
\item{url}{A URL to check}

\item{msg}{A message that contains \%s to replace in the site name}

\item{error}{Throw an error if the site is down; otherwise return a logical}
}
\value{
logical
}
\description{
Check if site is available
}
\keyword{internal}
