% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{stats}
\alias{stats}
\title{Check Stats}
\usage{
stats(text, ...)
}
\arguments{
\item{text}{the search table (or list of scienceverse objects)}

\item{...}{arguments to pass to statcheck()}
}
\value{
a table of statistics
}
\description{
Check Stats
}
\examples{
filename <- system.file("grobid", "incest.xml", package="papercheck")
study <- read_grobid(filename)
stats(study)
}
