% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osf.R
\name{summarize_contents}
\alias{summarize_contents}
\title{Summarize Directory Contents}
\usage{
summarize_contents(contents)
}
\arguments{
\item{contents}{a table with columns name, path such as from \code{osf_contents()}}
}
\value{
the table with new columns is_readme, is_data, is_code, and best_guess
}
\description{
Summarize Directory Contents
}
