% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/validate.R
\name{validate}
\alias{validate}
\title{Validate a module}
\usage{
validate(paper, module, ...)
}
\arguments{
\item{paper}{a paper object or a list of paper objects}

\item{module}{The name of a module or path to a module}

\item{...}{further arguments to match to the module output}
}
\value{
A validation list object
}
\description{
Validate a module
}
