% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{concat_tables}
\alias{concat_tables}
\title{Concatenate tables}
\usage{
concat_tables(papers, name_path)
}
\arguments{
\item{papers}{a list of scienceverse study objects}

\item{name_path}{a vector of names that get you to the table}
}
\value{
a merged table
}
\description{
Concatenate tables across a list of scienceverse objects
}
\examples{
grobid_dir <- system.file("grobid", package = "papercheck")
papers <- read_grobid(grobid_dir)
references <- concat_tables(papers, c("refs", "references"))
}
