% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/module.R
\name{module_run_text}
\alias{module_run_text}
\title{Run text module}
\usage{
module_run_text(paper, args)
}
\arguments{
\item{paper}{the scienceverse paper object (or list of objects)}

\item{args}{a list of arguments to \code{search_text()}}
}
\value{
data frame
}
\description{
Run text module
}
\keyword{internal}
