% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/osf.R
\name{osf_retrieve}
\alias{osf_retrieve}
\title{Retrieve info from the OSF by ID}
\usage{
osf_retrieve(id)
}
\arguments{
\item{id}{an OSF ID or URL}
}
\value{
a data frame of information
}
\description{
Retrieve info from the OSF by ID
}
