% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_text.R
\name{search_text}
\alias{search_text}
\title{Search text}
\usage{
search_text(
  paper,
  pattern = ".*",
  section = NULL,
  return = c("sentence", "paragraph", "section", "match"),
  ignore.case = TRUE,
  ...
)
}
\arguments{
\item{paper}{a paper object or a list of paper objects}

\item{pattern}{the regex pattern to search for}

\item{section}{the section(s) to search in}

\item{return}{the kind of text to return, the full sentence, paragraph, or section that the text is in, or just the (regex) match}

\item{ignore.case}{whether to ignore case when text searching}

\item{...}{additional arguments to pass to \code{grep()} and \code{regexpr()}, such as \code{fixed = TRUE}}
}
\value{
a data frame of matches
}
\description{
Search the text of a paper or list of paper objects. Also works on the table results of a \code{search_text()} call.
}
\examples{
filename <- demoxml()
paper <- read_grobid(filename)

search_text(paper, "p\\\\s*(=|<)\\\\s*[0-9\\\\.]+", return = "match")
}
