% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search_text.R
\name{search_text}
\alias{search_text}
\title{Search text}
\usage{
search_text(
  paper,
  pattern = ".*",
  section = NULL,
  return = c("sentence", "paragraph", "div", "section", "match", "id"),
  ignore.case = TRUE,
  fixed = FALSE,
  perl = FALSE
)
}
\arguments{
\item{paper}{a paper object or a list of paper objects}

\item{pattern}{the regex pattern to search for}

\item{section}{the section(s) to search in}

\item{return}{the kind of text to return, the full sentence, paragraph, div, or section that the text is in, or just the (regex) match, or all body text for a paper (id)}

\item{ignore.case}{whether to ignore case when text searching}

\item{fixed}{logical. If TRUE, pattern is a string to be matched as is. Overrides all conflicting arguments.}

\item{perl}{logical. Should Perl-compatible regexps be used?}
}
\value{
a data frame of matches
}
\description{
Search the text of a paper or list of paper objects. Also works on the table results of a \code{search_text()} call.
}
\examples{
filename <- demoxml()
paper <- read_grobid(filename)

search_text(paper, "p\\\\s*(=|<)\\\\s*[0-9\\\\.]+", return = "match")
}
