% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stats.R
\name{stats}
\alias{stats}
\title{Check Stats}
\usage{
stats(text, ...)
}
\arguments{
\item{text}{the search table (or list of scienceverse paper objects)}

\item{...}{arguments to pass to statcheck()}
}
\value{
a table of statistics
}
\description{
Check Stats
}
\examples{
filename <- demoxml()
papers <- read_grobid(filename)
stats(papers)
}
